fixed CSS to work with Media Queries
I set it up to show different backgrounds depending on the media querie for Phone or Mac.  Then fixed the media queries so they will work across browsers.  There may be a centering issue on some of the stuff depending on the browser, but the issue with blocking the inputs should be fixed.